+++ framework/runtime/src/runtime/runtime.ts
@@ -107,7 +107,6 @@ export class Runtime<Agents extends { [name: string]: Runnable } = {}, State ext
     OrderedRecord.push(this.runnableDefinitions, ...OrderedRecord.toArray(options.projectDefinition?.runnables));
 
     this.container.register(TYPES.context, { useValue: this });
-    this.container.register(TYPES.llmModelConfiguration, { useFactory: () => this.config.llmModel || {} });
     this.container.register('pipeline_agent', { useClass: PipelineAgent });
     this.container.register('llm_agent', { useClass: LLMAgent });
     this.container.register('function_agent', { useClass: FunctionAgent });
@@ -133,11 +132,9 @@ export class Runtime<Agents extends { [name: string]: Runnable } = {}, State ext
 
   state: State;
 
-  readonly agents: Agents = new Proxy({}, { get: (_, prop) => this.resolveSync(prop.toString()) }) as Agents;
+  agents: Agents = new Proxy({}, { get: (_, prop) => this.resolveSync(prop.toString()) }) as Agents;
 
   private container: DependencyContainer = container.createChildContainer();
-
-  private runnables: OrderedRecord<Runnable> = OrderedRecord.fromArray([]);
 
   private runnableDefinitions: OrderedRecord<RunnableDefinition> = OrderedRecord.fromArray([]);
 
@@ -146,21 +143,18 @@ export class Runtime<Agents extends { [name: string]: Runnable } = {}, State ext
     else this.container.register(token, { useValue: dependency });
   }
 
-  private resolveSync<T extends Runnable>(id: string | RunnableDefinition): T {
-    if (typeof id === 'string') {
-      const runnable = this.runnables[id];
-      if (runnable) return runnable as T;
-    }
-
-    const definition = typeof id === 'string' ? this.runnableDefinitions[id] : id;
+  private resolveSync<T extends Runnable>(id: string | RunnableDefinition | T): T {
+    const definition =
+      typeof id === 'string'
+        ? this.runnableDefinitions[id]
+        : id instanceof Runnable
+          ? this.runnableDefinitions[id.id]
+          : id;
 
     if (definition) {
       const childContainer = this.container.createChildContainer().register(TYPES.definition, { useValue: definition });
-
       const result = childContainer.resolve<T>(definition.type);
-
       childContainer.dispose();
-
       return result;
     }
 
@@ -175,12 +169,14 @@ export class Runtime<Agents extends { [name: string]: Runnable } = {}, State ext
 
   register<R extends Array<RunnableDefinition | Runnable> = []>(...runnables: R): void {
     for (const runnable of runnables) {
-      if (runnable instanceof Runnable) OrderedRecord.pushOrUpdate(this.runnables, runnable);
-      else OrderedRecord.pushOrUpdate(this.runnableDefinitions, runnable);
+      OrderedRecord.pushOrUpdate(
+        this.runnableDefinitions,
+        runnable instanceof Runnable ? runnable.definition : runnable
+      );
     }
   }
 
-  async resolve<T extends Runnable>(id: string | RunnableDefinition): Promise<T> {
+  async resolve<T extends Runnable>(id: string | RunnableDefinition | T): Promise<T> {
     return this.resolveSync<T>(id);
   }
 
@@ -189,10 +185,19 @@ export class Runtime<Agents extends { [name: string]: Runnable } = {}, State ext
   }
 
   copy<State extends ContextState = ContextState>(
-    options: Required<Pick<RuntimeOptions<Agents, State>, 'state'>>
+    options: Pick<RuntimeOptions<Agents, State>, 'state' | 'config'>
   ): Runtime<Agents, State> {
     const clone: Runtime<Agents, State> = Object.assign(Object.create(Object.getPrototypeOf(this)), this);
-    clone.state = options.state;
+
+    // override state/config
+    if (options.state) clone.state = options.state;
+    if (options.config) clone.config = options.config;
+
+    // override container & register context with cloned runtime
+    clone.container = this.container.createChildContainer();
+    clone.container.register(TYPES.context, { useValue: clone });
+
+    clone.agents = new Proxy({}, { get: (_, prop) => clone.resolveSync(prop.toString()) }) as Agents;
 
     return clone;
   }